Chore: Sync fuzzer: Test with larger items, more initial items
