+++ packages/tools/fuzzer/types.ts
@@ -46,9 +46,11 @@ export const assertIsFolder: (item: TreeItem)=> asserts item is FolderRecord = i
 export interface FuzzContext {
 	serverUrl: string;
 	isJoplinCloud: boolean;
+	keepAccounts: boolean;
 	baseDir: string;
 	execApi: (method: HttpMethod, route: string, debugAction: Json)=> Promise<Json>;
 	randInt: (low: number, high: number)=> number;
+	randomFrom: <T> (data: T[])=> T;
 }
 
 export interface RandomFolderOptions {
